Feat/#58/post 도메인 개발 (#65)
* rename: 클래스명 리네이밍

* rename: 클래스명 리네이밍

* rename: 파일이동

* rename: 클래스명 변경

* refactor: PostStatus 삭제

* feat: 게시물 작성 서비스 코드 구현

* feat: Post 영속성 매퍼 구현

* refactor: PricePersistenceMapper 리팩토링

* feat: Post 저장 영속성 어댑터 구현

* refactor: PostImageEntity {post_no, image_order} unique 제약조건 설정

* feat: 게시물 작성 웹 어댑터 구현

* test: 게시물 작성 통합 테스트

* refactor: Price 정보 수정/삭제 시 자격 검증 로직 추가

* rename: Expert Out Port 클래스명 변경

* feat: 게시물 수정 서비스 구현

* feat: Post 조회 영속성 어댑터 구현

* feat: Post 수정 영속성 어댑터 구현

* feat: Post 수정 웹 어댑터 구현

* test: 게시물 수정 통합 테스트

* feat: 게시물 삭제 서비스 로직 구현

* feat: PostEntity, PostImageEntity 삭제 영속성 어댑터 구현

* refactor: 게시물 삭제 후 Expert의 활동 수, 최근 활동일 업데이트 로직 추가

* feat: 최신 PostEntity의 수정날짜를 조회하는 영속성 어댑터 구현

* rename: ExpertCommandPort 메서드명 변경

* rename: 컨트롤러 디렉토리 구조 변경 (adapter/in/web -> adapter/in)

* rename: domain 디렉토리 model 디렉토리 제거

* rename: in port 클래스명 리팩토링

- XxxUseCase로 통일
- 기본적으로 기능에 맞게 동사 + 목적어
- 만약 단순 CRUD인 경우:
    - Load → load()
    - Create → create()
    - Update → update
    - Delete → delete()
- controller
    - in port 따라가기

* rename: out port 클래스명 리팩토링

- 접미사 Port
- 클래스명
    - Create - create()
    - Update - update()
    - Read - findByXXX(), findByXXXOrderByYYY, existsByXXX()
    - Delete - delete()

* rename: Get-> Load

* refactor: API URL @RequestMapping으로 공통처리

* feat: 게시물 삭제 웹 어댑터 구현

* test: 게시물 삭제 통합테스트

* refactor: 게시물 저장 시 Expert의 활동 수, 최근 활도일 갱신 로직 추가

* test: 게시물 수정시 전문가 최근 활동일 갱신되는 부분 테스트 추가

* fix: 토큰 인증으로 인한 테스트 실패 문제 해결

* feat: 특정 전문가의 갤러리를 조회하는 서비스 로직 구현

* feat: 특정 expertNo를 가지고 isPinned가 True인 Post를 조회하는 영속성 어댑터 구현

* feat: 갤러리 조회 요청 영속성 어댑터 구현

* test: 특정 전문가의 갤러리 조회 통합 테스트

* rename: 패키지 이동

* feat: 전문가 갤러리 수정 서비스 로직 구현

* feat: PostEntity만 수정하는 영속성 어댑터 구현

* feat: 갤러리 수정(고정처리할 게시물 변경) 요청 웹 어댑터 구현

* test: 갤러리 수정(고정처리할 게시물 변경) 통합 테스트

* fix: 오타 수정

* rename: 메서드명 수정(XxxLimited -> TopN)

* feat: 게시물 검색어 추천 서비스 로직 구현

* rename: AppReq -> Command/Query / AppResp -> Result

* rename: Out Port 동사+명사 -> 명사+동사

* refactor: Project, Skill, Studio, Portfolio를 별도의 DTO 객체로 변경

* refactor: 전문가 승인 요청 API 입력값 검증

* rename: 클래스명 리팩토링

* fix: 충돌 해결

+++ core/src/main/java/com/picus/core/price/adapter/out/persistence/PricePersistenceAdapter.java
@@ -12,12 +12,14 @@ import com.picus.core.price.adapter.out.persistence.repository.OptionJpaReposito
 import com.picus.core.price.adapter.out.persistence.repository.PackageJpaRepository;
 import com.picus.core.price.adapter.out.persistence.repository.PriceJpaRepository;
 import com.picus.core.price.adapter.out.persistence.repository.PriceReferenceImageJpaRepository;
-import com.picus.core.price.application.port.out.PriceCommandPort;
-import com.picus.core.price.application.port.out.PriceQueryPort;
-import com.picus.core.price.domain.model.Option;
-import com.picus.core.price.domain.model.Package;
-import com.picus.core.price.domain.model.Price;
-import com.picus.core.price.domain.model.PriceReferenceImage;
+import com.picus.core.price.application.port.out.PriceCreatePort;
+import com.picus.core.price.application.port.out.PriceDeletePort;
+import com.picus.core.price.application.port.out.PriceReadPort;
+import com.picus.core.price.application.port.out.PriceUpdatePort;
+import com.picus.core.price.domain.Option;
+import com.picus.core.price.domain.Package;
+import com.picus.core.price.domain.Price;
+import com.picus.core.price.domain.PriceReferenceImage;
 import com.picus.core.shared.annotation.PersistenceAdapter;
 import com.picus.core.shared.exception.RestApiException;
 import lombok.RequiredArgsConstructor;
@@ -29,7 +31,7 @@ import static com.picus.core.shared.exception.code.status.GlobalErrorStatus._NOT
 
 @PersistenceAdapter
 @RequiredArgsConstructor
-public class PricePersistenceAdapter implements PriceQueryPort, PriceCommandPort {
+public class PricePersistenceAdapter implements PriceCreatePort, PriceReadPort, PriceUpdatePort, PriceDeletePort {
 
     private final PriceJpaRepository priceJpaRepository;
     private final PackageJpaRepository packageJpaRepository;
@@ -74,7 +76,7 @@ public class PricePersistenceAdapter implements PriceQueryPort, PriceCommandPort
     }
 
     @Override
-    public Price save(Price price, String expertNo) {
+    public Price create(Price price, String expertNo) {
 
         // PriceEntity 저장
         PriceEntity savedPriceEntity = savePriceEntity(price, expertNo);
@@ -142,8 +144,7 @@ public class PricePersistenceAdapter implements PriceQueryPort, PriceCommandPort
     }
 
     private PriceEntity savePriceEntity(Price price, String expertNo) {
-        PriceEntity entity = pricePersistenceMapper.toEntity(price);
-        entity.bindExpertNo(expertNo);
+        PriceEntity entity = pricePersistenceMapper.toEntity(price, expertNo);
         return priceJpaRepository.save(entity);
     }
 
@@ -207,14 +208,16 @@ public class PricePersistenceAdapter implements PriceQueryPort, PriceCommandPort
 
     private void updatePriceReferenceImageEntities(Price price, List<String> deletedPriceRefImageNos, PriceEntity priceEntity) {
         // 삭제
-        priceReferenceImageJpaRepository.deleteByPriceReferenceImageNoIn(deletedPriceRefImageNos);
+        if(!deletedPriceRefImageNos.isEmpty()) {
+            priceReferenceImageJpaRepository.deleteByPriceReferenceImageNoIn(deletedPriceRefImageNos);
 
-        // 삭제 후 이미지 순서 재정렬
-        List<PriceReferenceImageEntity> images =
-                priceReferenceImageJpaRepository.findAllByPriceEntity_PriceNoOrderByImageOrder(priceEntity.getPriceNo());
+            // 삭제 후 이미지 순서 재정렬
+            List<PriceReferenceImageEntity> images =
+                    priceReferenceImageJpaRepository.findAllByPriceEntity_PriceNoOrderByImageOrder(priceEntity.getPriceNo());
 
-        for (int i = 0; i < images.size(); i++) {
-            images.get(i).assignImageOrder(i + 1);
+            for (int i = 0; i < images.size(); i++) {
+                images.get(i).assignImageOrder(i + 1);
+            }
         }
 
         // 추가/수정
